requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.8
oauthlib==3.3.1
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
tweepy==4.16.0
urllib3==2.6.3